sysext.just: Support setting up services to uphold
System services that are included in sysexts are not started on boot,
even if enabled.

This works around this issue by using the `Upholds` directive [1] to
make sure that those units are started.

Note that this bypasses the enabled/disabled state, but it is still
possible to mask the unit, or remove the sysext entirely.

This will also set the `EXTENSION_RELOAD_MANAGER=1` option for the
sysexts which tell systemd-sysext to ask systemd to reload its
configuration when the sysexts are loaded [2].

[1]: https://www.freedesktop.org/software/systemd/man/latest/systemd.unit.html#Upholds=
[2]: https://www.freedesktop.org/software/systemd/man/latest/systemd-sysext.html

See: https://github.com/flatcar/sysext-bakery/blob/87f4af7e579846c6d9c16d7db70f3cfa1b57a805/create_kubernetes_sysext.sh#L89-L90